json status updates
If brimcap analyze or load is not writing to stdout (load never is), and if
stdout is not a terminal (or the -json flag is enabled) write json status
updates to stderr.

The json status messages are mostly the same as the status updates sent back by
the zqd pcap post endpoint; this should be an easy transition for the when the
brim desktop application starts calling brimcap directly.