Add brimcap launch command (#17)
Add the brimcap launch command. brimcap launch takes as argument the time span,
protocol, source ip and port, and destination ip and port describing
a connection and searches through all the indexed pcaps in the BRIMCAP_ROOT
packets matching the description. If the search returns a result the packets
are opened in wireshark.

Additional changes:
- add brimcap search: the same thing as brimcap launch except it writes the
  search result as a pcap file.
- brimcap index -root: If the -root flag is provided brimcap index will create
  a index of the pcap in the BRIMCAP_ROOT.