Zed update through "colorize zson" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2621, authored by @mccanne,
has been merged.

colorize zson

This commit add colorization of zson -Z text.  This is a simple start
that we can improve upon over time.  We should probably unify the
colors with the app (though the terminal color codes are limited and
aren't generic RGB values).  Type decorators are toned down in a
light gray to make them less obstrusive.  We also added -color to
outputflags so color can be disabled with -color=false and the
upcoming defaults settings

Closes brimdata/zed#2578