Zed update through "last overhaul of ZJSON" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2526, authored by @mccanne,
has been merged.

last overhaul of ZJSON

This commit implements one last overhaul to the ZJSON spec
to bring it into alignment with where the Zed design landed.
Importantly, types and type values are consistently encoded
using JSON structure so that clients can have an easier time
parsing and managing the Zed type context for a stream of
ZJSON records.  This also means that type-values are encoded
with JSON structure instead of using ZSON type strings that
in turn needed to be parsed.

This change follows the compiler AST pattern for Zed types
but keeps the simple encoding of values based on nested
JSON arrays of strings.  The types are now more verbose but
easier to parse at the client.  Also, the aliases hack was
removed in favor of Zed's model of embedded typedefs.

We also eliminated the use of package joe and removed it
as this was the last remaining vestige of joe-ness.

Closes brimdata/zed#1780
Closes brimdata/zed#2185
Closes brimdata/zed#2376
Closes brimdata/zed#2525