Zed update through "fix type-value consistency bug" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2718, authored by @mccanne,
has been merged.

fix type-value consistency bug

This commit fixes a bug where type values were created both with the
ZSON() method on zng.Type as well as zson.FormatType().  The latter
is correct as the ZSON() method does not create canonical form and
import cycles prevent an easy fix.  The workaround for now is to
move NewTypeType() into package zson and have it call zson.FormatType()
instead of zng.Type.ZSON().