Zed update through "Update page users see when connecting from a browser" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2722, authored by @philrz,
has been merged.

Update page users see when connecting from a browser

While updating the Troubleshooting article in the Brim wiki, I noticed that this page still mentions `zqd` and therefore should be updated.

Just to refresh folks on why this page exists, we once had Support issues back in the day because users on Linux didn't realize Brim was a desktop app and therefore they installed via `yum` or `apt` from the shell and then didn't know how to launch it. They poked around and saw the listening `zqd` process and figured they should connect to it from a browser.

The new page:

---

![](https://user-images.githubusercontent.com/5934157/118295608-ebd7a280-b490-11eb-96c9-31e66bcce469.png)