Zed update through "Simplify keyword search" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2692, authored by @mccanne,
has been merged.

Simplify keyword search

This commit simplifies the keyword search experience by eliminating the field shortcuts, replacing "=" for assignment (":=") and comparison ("=="),  and introducing a "matches" keyword for regex match and glob match. The `*=value` and `value in *` patterns are no longer available but will be subsumed by better syntax in a forthcoming PR.

This change will break the app so we need to port and do integration testing before merging this PR.