Zed update through "fix parallelizer bug with summarize key expressions" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2755, authored by @mccanne,
has been merged.

fix parallelizer bug with summarize key expressions

This commit fixes a bug in the flowgraph parallelizer where
the partials-in aggregator uses the wrong key expressions,
i.e., it uses the group-by key expressions "as is" rather than
referencing the key value already computed upstream.

This bug dates back to the original implementation of the
optimizer.

Closes brimdata/zed#2728