Zed update through "fix dataflow through cut operators" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2760, authored by @mccanne,
has been merged.

fix dataflow through cut operators

This commit fixes a problem where a query would be incorrectly
parallelized by merging on the wrong key because sufficient
dataflow info was not properly tracked.  While we fixed this
we broke another optimization that doesn't seem important to
maintain (and anyway was based on flawed logic).  These tests
are marked "skip".  When we implement proper dataflow analysis
as part of issue brimdata/zed#2756, we can turn the tests back on.

Closes brimdata/zed#2663