Zed update through "Reintroduce zapi command" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2774, authored by @mattnibs,
has been merged.

Reintroduce zapi command

Reintroduce a new version of the zapi command the uses the exact same
Commands that the zed lake command uses. The idea is to have functional
parity between zapi and zed lake.