Zed update through "Clarify Zeek docs, including description of reference shaper config" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2782, authored by @philrz,
has been merged.

Clarify Zeek docs, including description of reference shaper config

Since brimdata/zed#2296 removed the legacy `types.json` approach for restoring rich data types to NDJSON data, this PR now documents the new & improved Zed shaper approach to addressing this. While approaching the revision, I recognized that the scope of our Zeek docs has widened to the point where it justifies a brief leading README with links to separate detailed docs on multiple topics, with shaping being just one of them.

Since the platform is in transition and is expected to keep evolving, I've taken the approach here of being transparent about where current approaches are likely to be improved as we continue adding the enhancement from our backlog, linking directly to the open issues where appropriate. In addition to assuring users that sub-optimal workflows won't remain that way forever, I figure ambitious folks might even be inspired to chime in on issues or speak up on Slack and help us determine the priority of which ones we add first and/or how we implement them.

I'm aware that the ideal approach will eventually be to have a single, generic shaping doc and just link to it as a reference. However, since the rewrite of the docs with a generic data set is still a work-in-progress, for now I'm taking the approach of focusing on the Zeek-centric reference example as a way to introduce shaping, since it's expected to remain our most popular log format for MVP0 and a while after. But I _do_ want this to change over time.

Closes brimdata/zed#2594