Zed update through "encode type values in binary and move type context into zng" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2826, authored by @mccanne,
has been merged.

encode type values in binary and move type context into zng

This commit reworks the type value specification and implementation
so that ZNG-serialized type values have their own encoding independent
of the ZSON format.  This breaks the dependencing between zng and zson
and will allow the zng package to be moved to the top-level zed package.

This introduces an backward-incompatible change to the ZNG format as
type values are now encoded differently.  However, this change should
be minimally impactful as type values appear more often dynamically as
the result of a query and not so much in origin data.