Refactor analyze (#110)
The old way analyze was written was prone to hard to reproduce / hard to
debug concurrency bugs. Rewrite analyze to simplify things. This commit
divorces the analyze process command phase and the log tailer phase
which were previously unified. This arrangement makes it easier to
isolate the two phases of analyze which makes it easier to diagnose issues.

This refactor fixes the deadlock bug encountered when one analyzer process
would error out in the midst of analysis.

Also:
- Have the load commit use the add / commit endpoints instead of the
  deprecated log post endpoints.

Closes #109
Closes #71