Zed update through "fix bug in zng type ID decoding" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2847, authored by @mccanne,
has been merged.

fix bug in zng type ID decoding

This commit fixes a bug where escaped type IDs were not properly
decoded in the zng reader.  The fix was to add the escaped base
value to the decoded delta.

Fixes brimdata/zed#2846