Zed update through "Tweaks to top-level language README" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2878, authored by @philrz,
has been merged.

Tweaks to top-level language README

Now that brimdata/zed#2866 has been merged, here's the first of what I expect to be several more passes through the docs with iterative improvements.

Like you may have seen in other docs, one of the things I'm fussy about is linking to open issues if we're showing syntax that doesn't actually work that way in the tools yet. While it makes the text a little less pretty, I prefer that to the thought of a user attempting to use the syntax we show and wondering why it doesn't work. We can let the desire for prettiness be our motivation to address the issues. :smile:   I'll be putting comments in the issues as a reminder to go back and remove these links as we address them.