Zed update through "store config state in a journal and improve index object model" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2931, authored by @mccanne,
has been merged.

store config state in a journal and improve index object model

This commit changes the storage model for pool/lake configuration
from a json file to a transaction journal.  It also improves the
index abstraction into an index-rules configuration stored as a
journal.

We also updated the lake design docs to reflect this new model
and change reference to data "segment" to "data object".

The pool rename command has been added.

Closes brimdata/zed#2547