Zed update through "csvio.Reader: Remove extraneous quotes" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2942, authored by @mattnibs,
has been merged.

csvio.Reader: Remove extraneous quotes

Add a reader on top of the go csv reader that looks for fields where
quotes do not cover the entirety of a field. If such a case is found
the quotes are stripped from the field.

For instance, the line:
field1,"field2" extra
Would get converted into:
field1,field2 extra

Closes brimdata/zed#2862