Zed update through "Limit scope of reserved words" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2985, authored by @mattnibs,
has been merged.

Limit scope of reserved words

Allow IDs to have the name of reserved words (e.g. by, from, where, etc).

In order to make this possible change Where clauses to accept a
LogicalOrExpr instead of BooleanSearch as it was previously doing. The
BooleanSearch made it impossible to distinguish between a keyword such
as by and limit and an ID search primitive.

Closes brimdata/zed#2398