Zed update through "improve the branches model to be more git-like" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2994, authored by @mccanne,
has been merged.

improve the branches model to be more git-like

This commit improves upon the recent prototype of branching for the
zed lake.  The new model is very much git-like, where commit objects
are pool-wide in scope and stored independently of the branch structure.

Branches now are simply named pointers to commit objects. Just like git.

The commit objects of a pool form a tree rooted at the null commit
and pointing backward from most recent object to oldest.

Scanning can be performed at any commit ID.  There need not be a
branch reference.  Obviously, this provides a form of time travel,
but time-based time travel will be soon implemented by the traveling
backward in the branches and pools tables so that the pool and branch
state can be accessed at any point in time.

All of the marshaling has been changed to use "package style"
types as this will avoid collision in naming.  (In the future,
we can also manging the Zed time names separately from the
Go struct names as you can bind by an arbitrary name.)

Branch merging is done by transactionally and writers can concurrently
commit to either branch while the merge is running.  Rebase will be
implemented in a future PR.  There is no way yet to move a branch
pointer (other deleting it and recreating it).  A rebase will allow
live ingest to a child branch while periodic merges may occur to a
parent then rebasing along with period rebasing of the child to the
tip of the parent.  There is no need to synchronize the merge and
the rebase as they will both be transactionally consistent. (This will
be implemented by rewriting any commit objects in the child past
its merge point in a optimistic fashion until a successful transaction
is written.)

The meta-query syntax is updated from pool/branch[meta] to
pool@brach:meta or pool@commit:meta.

We also added a "rawlog" meta-query to provide access to the
underyling pool and branch config journals (rather than just
getting a snapshot of the tip of the config values).

The lake branch command was added as a preferred way of creating
and deleting branches instead of using the lake create command
(which is no used only for pools).

There is a new "undo" command to revert a commit.  The delete
command is restricted to deleting individual objects and is
no longer overloaded to "delete" a commit.  (There were some
bugs in that logic anyway.)

Now that we have branches, we added branch annotations to the
zed lake log command so you can see here which branches point to
which commit objects.