Zed update through "Fixes to Zed lake design doc after following as a user" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3014, authored by @philrz,
has been merged.

Fixes to Zed lake design doc after following as a user

After the merge of brimdata/zed#2994, I did another pass through the Zed lake design doc, reading it as a user would and following along with the example commands. I've proposed a set of corrections that I think are appropriate. I'll add some comments to draw attention to where I think additional input is justified.