Zed update through "add lake checkout command and rework CLI args to use HEAD" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3020, authored by @mccanne,
has been merged.

add lake checkout command and rework CLI args to use HEAD

This commit adds some more fit and finish to our adaptation of the
Git design pattern.  We added a checkout command and the notion
of HEAD that is tracked in ./.zed_head.  The -p flag has been
eliminated.

The checkout command and the HEAD file provide a convenience and
is not required as all commands have a "-head" option at the lake
level.  We envisage that checkout would be used in interactive
workflows and automation and orchestration tooling would hardwire
their use of -head.

You can now query a lake without an implicit "from" operator,
in which case, HEAD is assumed.  You can also use "from HEAD".

We also changed the output flag -d to -D to make room for the -d
flag in the branch command.

The lakeio printer has been updated to output the familiar HEAD
notation in the log (e.g., "HEAD -> branch").

Finally, we created a new package called lakeparse to hold the
lake-oriented parsing functions that are used by the compiler,
the CLI, and the lake.  ParseID() has finally found its true home.

With a little more massaging, we should be able to checkout a
detached HEAD with an arbitrary commit ID.  See issue brimdata/zed#3019