Zed update through "Zeek shaper updates for v4.1.0" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3021, authored by @philrz,
has been merged.

Zeek shaper updates for v4.1.0

For some time now, each time a new Zeek version is released, I use our [print-types.zeek](https://github.com/brimdata/zeek/blob/master/brim/print-types.zeek) script to see if there's new log types in the default output and/or new/changed fields in the existing log types. I then update the reference Zed shaper appropriately. Since this is data with which we've got a lot of prior experience, I feel like it's good shaping practice.

When I went to do this for the recently-released Zeek v4.1.0, for the first time since maintaining the reference shaper, I actually saw field names _change_ from what they were prior. This poses a unique challenge because when fields were only being added, the only side effect of applying the newest shaper on older logs was that there'd be some extra fields filled in with `null`. In fact, I've always thought this has an up side, since it "future proofs" the user for when they _do_ upgrade to the newer Zeek release, since their shaped Zeek logs would continue to fit within the type definition they're already been using. But to cover fields that change, it seemed like I'd either have to:

1. Start maintaining separate version-specific reference shapers. Specifically, that would now include one for Zeek v4.0.3 and earlier and another for Zeek v4.1.0 and newer. I'd maintain just those two until yet another set of field name changes comes along.

2. Continue to maintain a single shaper, but introduce flow control logic to detect the changes that are unique to Zeek v4.1.0 and newer and apply them only when necessary.

In the interest of maintaining minimal redundant config and showing an approach that might be applicable in other contexts, in this PR I'm opting for the latter approach.

The diffs on which the changes are based is in the attached [types-diff.txt](https://github.com/brimdata/zed/files/7117335/types-diff.txt).