Zed update through "change behavior of lake checkout state" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3024, authored by @mccanne,
has been merged.

change behavior of lake checkout state

This commit changes the default behavior of the checkout command
to store the HEAD pointer in $HOME.  The environment variable
ZED_HEAD_PATH can be defined to override this behavior.  In particular,
setting ZED_HEAD_PATH to ".", will revert to the previous behavior
where the HEAD pointer is stored in the current working directory.