Zed update through "move the lake design doc to docs/lake/README.md" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3034, authored by @mccanne,
has been merged.

move the lake design doc to docs/lake/README.md

Despite the large diffs here, this is just a move of the lake design doc to the lake README and the old README to a file called tutorial.md.  I also added a link from the top-level README the lake README/design doc.