Zed update through "properly format empty string field name" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3054, authored by @mccanne,
has been merged.

properly format empty string field name

The JSON spec allows empty string to be an object key and
it was ZSON's intention to do the same, but the ZSON formatter
was incorrectly printing out the empty field name without quotes.
This commit fixes this bug and adds a test.