Zed update through "improve onboarding for installing and trying zed" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3098, authored by @mccanne,
has been merged.

improve onboarding for installing and trying zed

This commit improves the zed README with a more succinct description
of how to get up and running. @garrisonhess spent 2 hours futzing
with our repo before he could get a python client hitting the Zed service.
His feedback led to these improvements.

The FAQ is coming soon in a subsequent PR.

Closes brimdata/zed#2570