Zed update through "better index format for multi-type multi-keys" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3141, authored by @mattnibs,
has been merged.

better index format for multi-type multi-keys

For microindexes, store the key name(s) in the trailer (without type)
and do not rename the key to "key", rather keep the original record
name.

Allow a key to have multiple typed values.

Closes brimdata/zed#2923