Zed update through "allow dot operator to work on union containing record" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3178, authored by @mccanne,
has been merged.

allow dot operator to work on union containing record

This commit arranges for the dot operator to work on union values when the underlying value is a record.  We need to generalize this across all operators in the system, where we automatically promote union values to their underlying value when used in expressions expecting a non-union type.