Zed update through "fix lake/ztests/delete.yaml" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3236, authored by @nwt,
has been merged.

fix lake/ztests/delete.yaml

This applies the change to service/ztests/delete.yaml in brimdata/zed#3227 to the
similar test in lake/ztests/delete.yaml.

The test in lake/ztests/delete.yaml is flaky because it expects
"from pool@branch:objects" to emit objects in creation order, but the
actual order is undefined.  Fix it by grabbing the first object ID
immediately after creation.