Zed update through "attempt to detect CSV input automatically" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3277, authored by @nwt,
has been merged.

attempt to detect CSV input automatically

In zio/anyio.NewReaderWithOpts, return a zio/csvio.Reader if no other
format matches, the first line of input contains a comma, and a call to
match succeeds.

For brimdata/zed#2517.