Zed update through "allow arbitrary top-level values in ZNG" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3301, authored by @nwt,
has been merged.

allow arbitrary top-level values in ZNG

Relax the zio/zngio.Reader requirement that top-level values must be
records.  Other parts of the system continue to enforce this requirement
for now.

Depends on brimdata/zed#3300.