Zed update through "Migrate all zed lake subcommands to zed" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3314, authored by @mattnibs,
has been merged.

Migrate all zed lake subcommands to zed

Make all the subcommands of zed lake, subcommands of zed and get rid of
`zed lake`. The pr also unifies zed api with zed lake- Zed lake services
can be operated on by setting the -lake argument to the URI of a Zed lake
service.