Zed update through "fix join when first left key value is non-null but zero-length" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3334, authored by @nwt,
has been merged.

fix join when first left key value is non-null but zero-length

The join operator ignores the first value encountered for the left join
key if its zed.Value.Bytes is non-null but zero-length.  Fix this by
preserving the nil-ness of that field when setting join.Proc.joinKey.