Zed update through "Enable streaming stats on query endpoint" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3372, authored by @mattnibs,
has been merged.

Enable streaming stats on query endpoint

Stats were not getting sent on a periodic basis for the query endpoint.
This was causing long running queries on a service to time-out when
served behind an AWS load balancer: The load balancer would see no data
getting transmitted and time-out the connection.

Add a time Ticker in the query handler and also flush the response data
when a control message is sent.

Closes brimdata/zed#3371