Zed update through "Change group by every syntax" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3451, authored by @mattnibs,
has been merged.

Change group by every syntax

Remove the current "every [duration]" syntax for the summarize proc and
instead add function "every" that can be used as a key in summarize.
This change also keeps intact the optimizations done on every queries by
looking for an every call in a summarize proc's keys.

Closes brimdata/zed#2213