Zed update through "add structured errors, change ZSON enum syntax" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3458, authored by @mccanne,
has been merged.

add structured errors, change ZSON enum syntax

This commit adds structured errors to Zed as part of the Dec 2021
changes to the Zed formats.  Since a parser ambiguity arose that was
solved by the new enum syntax, we rolled the enum changes into this PR
as well.

Also bstring and stringy errors are now gone, so we got rid of the IsStringy
pattern and replaced it everywhere with a compare to type string.

The diffs are somewhat verbose because the change requires that errors
are now allocated from a type context so many more entities now need a zctx.

Closes brimdata/zed#3320