Zed update through "change ZSON types to use angle brackets instead of parens" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3459, authored by @mccanne,
has been merged.

change ZSON types to use angle brackets instead of parens

This commit updates the ZSON implementation to conform to the
Zed format changes of December 2022.  The Zed language had been
previously udpated but we were waiting for the big changes to
update the serialization format.