Zed update through "new version of ZNG" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3477, authored by @mccanne,
has been merged.

new version of ZNG

This commit updates the ZNG protocol specification and implementation
to conform to the December 2021 upadtes to Zed along with a better way
to frame payloads.  This new approach will make it easy to add
parallelism to the output compression path.

As part of the refactor, we create a synchronous ZNG scanner for
test and debug alongside the parallel scanner.  There is also a
new input flag (-threads) to specify the number of threads used
by the zng scanner (default of 0 uses go maxprocs and setting it
to 1 forces the synchronous scanner).

On no output, the ZNG writer emits now outputs nothing instead of a
single-char EOS.

We also added a new dev command zngdump that prints metadata
about serialized ZNG frames.  This will come in handy down the
road as we explore performance and how different types of ZNG
impact perf in different ways.

This also fixes the null value bug (echo "null(int64)" | zq - | zq -).

Co-authored-by: Noah Treuhaft <noah.treuhaft@gmail.com>