Zed update through "backward compat for ZNG prior to January 2022" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3481, authored by @mccanne,
has been merged.

backward compat for ZNG prior to January 2022

This commit adds a reader for the temporary format called "zng21",
which decodes the old ZNG format prior to the changes made in
January 2022.  The relevant pieces from older versions of
packages zngio, zcode, and zed were copied into zio/zng21io.

This code will be discareded probably within 12 months so we shouldn't
spend too much time perfecting it.