Zed update through "remove cut() function" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3515, authored by @mccanne,
has been merged.

remove cut() function

The cut() function was added to the Zed lanuage before it had
record expressions (and before it was called Zed). This commit
removes it and updates the tests to use record expressions.