Zed update through "fix bug compiler bug with regexp search in expressions" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3520, authored by @mccanne,
has been merged.

fix bug compiler bug with regexp search in expressions

The compiler was missing a case for regexp search predicates used
inside of an expression.  This commit provides the trivial fix
and adds a test.