Zed update through "add with clause to over" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3521, authored by @mccanne,
has been merged.

add with clause to over

This commit adds an experimental syntax to change the "let" prefix
to a "with" suffix in a scoped over subgraph.  For now, all syntax
options co-exist and we will clean up the grammar when we decide
on the final form.