Zed update through "collect: ensure correct selector for type" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3531, authored by @mattnibs,
has been merged.

collect: ensure correct selector for type

With canonical unions the collect aggregation was not deriving the
correct selector index for a union of values since the type order may be
rearranged when the zed.Context creates the union type. Fix this and add
a lookup table to zed.TypeUnion so it's easier to find a type's
selector.