Zed update through "simplify methods to zed.Value and allow method chaining" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3532, authored by @mccanne,
has been merged.

simplify methods to zed.Value and allow method chaining

This commit simplifies the use of zed.Values by allowing method
chaining.  In this approach, a missing value is modeled as nil
since the receiver does not have access to a type context to create
a missing error.  Instead a caller with a type context can check
for nil and create a missing error value if desired.

In debugging tests, we noticed that the old issue to allow values with
missing pool keys is moot so we simply removed the relevant check for
missing added test coverage and are closing it.

We are also closing brimdata/zed#1347 as half of this is addressed here and the other
half (constructing values using method chaining) is also moot now that we
have comprehensive marshaling.  This may come back later but is not
high priority right now.  This also address the intent of brimdata/zed#2166.

In a subsequent PR, we will merge recordval.go into value.go and delete
the former.  For now, the diffs are easier to review on this commit.

Closes brimdata/zed#2637
Closes brimdata/zed#1347
Closes brimdata/zed#2166