Zed update through "fix bug in data object writer" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3536, authored by @mccanne,
has been merged.

fix bug in data object writer

This commit fixes a bug where the span range of the object was
clobbered because the zed.Value from the object's records would
be freed while the object data structure was held across objects.
The fix is to copy the zed.Values of the span.