Zed update through "fix buffer pool for uncompressed buffers" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3540, authored by @mccanne,
has been merged.

fix buffer pool for uncompressed buffers

This quick commit speeds up zq searches by 25% by sending the
decompressed buffers to a pool along with the compressed buffers.
It's a bit hacky and this could all use some more work but
I think it's okay to check in for now.