Zed update through "add error function" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3544, authored by @mccanne,
has been merged.

add error function

This commit adds an error() function to the Zed language to wrap
any Zed value as an error.  We also changed the name of iserr()
to is_error() to be consistent with other snakecase names.