Zed update through "make default LZ buffer size much bigger" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3545, authored by @mccanne,
has been merged.

make default LZ buffer size much bigger

This commit changes the default LZ buffer size in the ZNG writer
from 16KB to 500KB leading to a speedup of 5X (in combination
with the checked-in buffering improvements) for Boyer-Moore string
search.  The compressed ZNG files are materially smaller.

I tried 1MB too (after fixing the bug in brimdata/zed#3543) but it didn't seem
to make a difference over 500KB.