Zed update through "speed up zio/jsonio.Reader with a builder" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3558, authored by @nwt,
has been merged.

speed up zio/jsonio.Reader with a builder

zio/jsonio.Reader is slow for a few reasons.

1. To normalize record field order, it unmarshals JSON to Go values and
   then marshals back to JSON (via encoding/json).

2. To unmarshal JSON to Zed values, it uses a zio/zsonio.Reader, which
   is not optimized.

3. Each call to its Read method calls zsonio.NewReader, which does a
   number of allocations.

Speed it up by replacing all of that with an optimized builder that
minimizes the number of allocations (though encoding/json.Decoder.Token
still does too many).

Note that unmarshaling via encoding/json can discard precision for
numbers because it unmarshals them into a float64.  Removing this step
accounts for the output change in zio/jsonio/ztests/reader.yaml.

Reading NDJSON with `-f json` is about three times faster for me with this branch compared to main.

Depends on brimdata/zed#3555.