Zed update through "emit missing errors in yield" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3569, authored by @mccanne,
has been merged.

emit missing errors in yield

Previously, the yield operator filtered missing errors and emitted
quiet errors.  This commit reverses this so quiet() works correctly
with yield (as was intended and documented in the quiet function).