Zed update through "format Zed type value as string in Parquet writer" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3592, authored by @nwt,
has been merged.

format Zed type value as string in Parquet writer

For a Zed type value field, zio/parquetio.Writer creates a column with
the Parquet STRING logical type but writes the binary type value to it.
This doesn't make much sense.  Format the type value with
zson.FormatTypeValue so we write an actual string instead.