Zed update through "preserve Zed record field order in JSON output" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3638, authored by @nwt,
has been merged.

preserve Zed record field order in JSON output

zio/jsonio.marshalRecord converts a Zed record to a Go map, which
json.Marshal then encodes as a JSON object with sorted keys.  Change the
map to a type that implements the json.Marshaler and preserves record
field order in JSON output.