Zed update through "fix zio/zngio.worker.run goroutine leak" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3655, authored by @nwt,
has been merged.

fix zio/zngio.worker.run goroutine leak

zio/zngio.worker.run expects work.resultCh to be buffered so that a send
will not block, but it isn't, causing a zngio.scanner to leak goroutines
if not read to completion.  Fix by making the channel buffered.