Zed update through "fix zio/zngio.scanner.Pull hang" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3656, authored by @nwt,
has been merged.

fix zio/zngio.scanner.Pull hang

zio/zngio.scanner.Pull expects that a receive from a result channel will
not block, but it will if context cancelation prevents the result
channel from making its way to a worker in scanner.start.  Close the
result channel when this happens so the receive will not block.