Zed update through "unions: Don't use selector for nulls" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3666, authored by @mattnibs,
has been merged.

unions: Don't use selector for nulls

Change union.SplitZNG so if the passed value is null (i.e. has no
selector) the returned type is the type of the union value.

This change has implications for unions in zst- the readers and writers
of which expected every value to have a selector. Add a presence
reader/writer to the union writer to handle null values in the union.