Zed update through "Make unflatten inverse of flatten" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3652, authored by @mattnibs,
has been merged.

Make unflatten inverse of flatten

Make unflatten inverse of flatten meaning the unflatten now accepts an
array of key value records and returns them into a single record. Rename
the former unflatten to zeek.