Zed update through "Simply ZSON formatting of mixed-type maps" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3684, authored by @mattnibs,
has been merged.

Simply ZSON formatting of mixed-type maps

Do not decorate the elements of a mixed-type map. Decorate a mixed
typed map if when the union type includes elements that are not in
the array or if elements are a named union.

Closes brimdata/zed#3400