Zed update through "fix flakiness in compiler/parser/ztests/sample.yaml" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3700, authored by @nwt,
has been merged.

fix flakiness in compiler/parser/ztests/sample.yaml

The test in compiler/parser/ztests/sample.yaml went flaky in brimdata/zed#3695
because the sort operator picks field s1 as the sort key if the first
value it sees is {s1:"a",s2:"b"}.  Fix by sorting on this.