Zed update through "handle non-record types in fuse operator" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3717, authored by @nwt,
has been merged.

handle non-record types in fuse operator

As mentioned in the comment in runtime/op/fuse/ztests/mixed.yaml, a
shaper bug prevents this from working correctly when the input contains
a mix of record and non-record types.