Zed update through "update tutorials" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3726, authored by @mccanne,
has been merged.

update tutorials

This commit updates the schools data tutorial to be consistent
with the rest of the docs.  Since the join example doesn't use
the schools data, we broke it out into its own doc.

The lake and zed-primer tutorials are coming in subsequent PRs.