Zed update through "make zed use command and -use flag more pool-centric" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3744, authored by @nwt,
has been merged.

make zed use command and -use flag more pool-centric

The zed use command and -use flag interpret their argument as a branch
or commit if it does not contain the @ character.  Change them to
interpret it as a pool with implied branch main, and remove the now
redundant zed use -p flag and the confusing -b flag.

Closes brimdata/zed#3243.

@mccanne: If this looks good, I'll suggest necessary updates on brimdata/zed#3735.