Zed update through "has_error function" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3724, authored by @mattnibs,
has been merged.

has_error function

Add has_error function which will recurse a zed value and return true if
the value or any of its leafs an error. has_error memoizes the types it
has seen so it won't have to recurse types more than once (unless the it
contains an error type in a union in which case it'll always have to
recurse).

Closes brimdata/zed#3719