Zed update through "move formats bullet in docs/README.md" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3770, authored by @mccanne,
has been merged.

move formats bullet in docs/README.md

This commit moves the formats bullet in docs/README.md to last
position so the bullets are new aligned in priority order for
a new user seeking docs on the system.