Zed update through "remove scaling from Zed duration and time conversions" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3809, authored by @nwt,
has been merged.

remove scaling from Zed duration and time conversions

Converting a Zed duration or time to an integer or floating point type
for a cast, coercion, or comparison scales from nanoseconds to seconds
and vice versa.  This scaling is surprising.  Remove it.

Closes brimdata/zed#1622.