Zed update through "more updates to online docs" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3815, authored by @mccanne,
has been merged.

more updates to online docs

This commit integrates the install description that was in
the top-level README with the docusaurus install and adds
a link from the README to the online docs.

We also created docs/libraries and moved the Python client
docs here and created simple initial docs for Go client access.
The Zealot library needs a bit more work so we have a
placeholder for the Javascript library docs.  We should
get to this soon and integrated observable notebooks as
reference use cases.