Zed update through "Disable markdown check for github link" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3836, authored by @mattnibs,
has been merged.

Disable markdown check for github link

The markdown check for a link to documentation on the github rest API
has recently been consistently returning 403 and therefore causing the
test to fail even though the link is working for an unauthenticated
browser client. Disable the check on the specific link for now.
The issue is being tracked here: https://github.com/github/feedback/discussions/14773