Zed update through "Proper coercion for aggregation math fns" by mattnibs
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/3826, authored by @mattnibs,
has been merged.

Proper coercion for aggregation math fns

Instead of using the first type to set the type for aggregation min,
max, and sum, use coercion rules to select the right type from the range
of values.